Added full trap functionality! Also, patrol monster works
Get the trap as a pickup (see prefabs) in the past
It is added to the inventory (middle item).
That item is a BUTTON. Move player to where you want to set the trap, and press the inventory button to locate it at the player position.
Switch to the future, and watch the patrol monster walk into it!